backend/requirements.txt
# Core
fastapi
uvicorn[standard]
pydantic
pydantic-settings
email-validator

# Database
sqlalchemy
asyncpg
alembic

# Security
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1

# Logging
loguru

# Utilities
python-multipart
python-magic
pillow
aiofiles
